requirements.txt

asgiref==3.8.1
    # via django

django==4.2.23
    # via
    #   -r requirements.in
    #   djangorestframework
django-environ==0.12.0
    # via -r requirements.in
djangorestframework==3.16.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
python-decouple==3.8
    # via -r requirements.in
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via django
aiogram~=3.21.0
sqlalchemy~=2.0.41
pika~=1.3.2
asyncio~=3.4.3
aio-pika~=9.5.5
django-debug-toolbar~=5.2.0
